+++ platform/ios/src/Mapbox.h
@@ -50,6 +50,7 @@ FOUNDATION_EXPORT MGL_EXPORT const unsigned char MapboxVersionString[];
 #import "MGLVectorSource.h"
 #import "MGLShapeSource.h"
 #import "MGLAbstractShapeSource.h"
+#import "MGLComputedShapeSource.h"
 #import "MGLRasterSource.h"
 #import "MGLTilePyramidOfflineRegion.h"
 #import "MGLTypes.h"